requirements.txt
flask>=2.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
img2pdf>=0.5.0
pillow>=10.0.0
werkzeug>=2.0.0
gunicorn>=21.0.0